Unify rate limiting logic
Implement the http request rate limiting in a universal way using a
universal GenServer and a Tesla middleware. Each service has its own
rate limiting server, which keeps track of the rate at which the service
is called. The rate limit check is done in Tesla middleware, which is
going to throttle the requests if we are above the rate limit.