Make the price alert logic execute parallel
This should provide much better performance when we have some new
projects for which we want to catch up with prices. The rate limiting
will ensure that we don't overload the CMC serves and the price alerts
won't stop happening. The alerts will just become a bit slower. Let's
test this approach and see how it perform compared to the old linear approach.