Add a private API for fetching the projects
Added an API which is behind basic auth for fetching the current
projects we track. Only projects which are successfully updated from CMC
will be returned. There will be more data coming out of this API as we
scrape more information about the projects.

Also separated the rate limiting for CMC into several separate buckets,
as we are making requests to different domains.