Import historical github activity of projects
Added a worker, which fetches the github archives for a particular date
and hour and computes the number of github events that happened for each
project we track during that hour.

The data is imported a time series database after that.

It should be possible to run this on many machines in parallel and
process the whole github archive within several hours.