Fix the etherscan response parsing
After a change in the Requests module, all the return values changed
from strings to maps, so the logic that depends on this should be
changed. I think we might need to just remove these modules and define
the structs in the Requests module, so that everything around the
requests could be in one place.