Schedule github activity scraping each hour
Add a gen server, which will schedule github activity scraping each
hour. The server checks if there are currently any jobs running and
won't schedule anything if it sees that there are jobs underway. This
logic needs further improvement, as it should check only the queue of
the github activity workers, but for now this information is not
provided by the INFO API in faktory.