Split the sanbase deployment into 3 different pods
We have one pod, which handles only web traffic and we want to have at
least 2 instances of this one for redundancy.

We have one pod for handling faktory jobs. These are going to scale
up/down depending on the amount of work scheduled in the queues.

We have one pod, which is only responsible for scraping prices, tickers,
scheduling github activity for scraping, etc. We aer going to have a
single instance of this, because of synchronization reasons.

In the long run we should split these responsibilities into different
repos with separate images. This will be possible when we extract our
common login into a separate hex package, which will be imported by all
these apps.