Optimize the github activity tracking and scheduling
Download the github archives in S3, so that when they need to be
reprocessed we can download them directly from there. The bandwidth to
S3 will be much bigger than to the external service.

Also improved the scheduling to keep track which archives are processed
for each project and do not reprocess archives if they are all scraped
before, even if there is no activity for them.