Improve the github activity scheduler
The current GH activity scheduler is timing out when trying to schedule
jobs. This is because of a very big query, which can't be optimized
without creating a gigantic index on it.

In this PR, the computations are broken down into many smaller queries
to the DB. Given that the scheduling is much faster than the processing,
even a bit slower scheduling will fine.