Add all the voting models and some basic logic
Here we add all the models around voting in the DB, together with some
basic logic for creating new poll, validating posts and votes. There are
some simple tests for the logic.

The next steps will be to add GQL apis for getting the voting data and
adding new posts and votes.