Add deletePost API and fix cascading deletes
Users can delete their posts now. When a post is deleted all the
attached votes will be deleted too.

Cacading deletes will happen for posts when deleting polls and users.

Also changed the currentPoll API to include the unapproved posts, as we
will filter these out on the client side.