Refactor ProjectCard component
Changes the ProjectCard component to show a small card in mobile view
initially and after the user clicks on the card it expands to show more
data.

Updates the ProjectsMobile component so the size of each card is
calculated automatically. This allows for having cards with different
sizes and also when the size of the card changes (when the user clicks
on the card) everything updates properly.