Use stored total_supply
Removes the multiplication to decimal places so we'll use the
total_supply in the base token instead of the smallest amount.

Fixes the scraper's selector for total_supply to be closer to the real
css selector.

Uses the total_supply stored in the Project as a fallback if
LatestCoinmarketcapData is missing the total_supply.