Use everywhere Sanbase Case templates. Make async: false explicit in all tests.
We currently do not have issues with the performance of the tests.
Making them all async has the benefit of making use of :shared mode of
the Ecto SQL Sandbox. If not in :shared mode the test will fail if the test process
spawns a new process (this can be hidden behind a call) which executes
an Ecto function. The :shared mode cannot run tests concurrently (async: true).