Market segments queries return proper JSON object
Market segments queries return a JSON object with name and count instead
of the string representation of the object.

The front end changes the list of categories based on the type of
projects that is selected and when the user changes the type of the
project the categories filter is cleared.