ExchangeInflow signal can trigger multiple times during the interval
If a signal is not triggered for the whole interval_days() interval it
won't be triggered for any subinterval. This allows us to optimize how
to recalculate the inflow for an interval that starts on the last
datetime the signal was triggered. That will allow us to report on a big
inflow when it happens despite that a previous big inflow triggered the
signal already.