Reimplement the Notification module
The previous implementation was used only for price-volume diff signal.
The issue was that it only needed to check if there was a notification
sent, but saved every notification that was sent. On stage this has
generated 200k rows of data. And it will only grow faster after more
than one signal is using it. Now it is rewritten so we only save the
last time a notification type has been sent for a given project