Change the max_concurrency and the time between cmc requests
Reduce the max_concurrency to 1 so when a 429 response is hit, the
message queue of the rate limiter does not have projects waiting to
execute that will hit that rate limit.
Decrease the time between requests from 1 second to 100 milliseconds to
balance the concurrency change.
These changes won't cause performance issues as the rate limit is hit
pretty easy and often, so the concurrency reduction brings more good
than harm