Introduce the SanbaseDataloader that unifies all dataloaders
The reason behind this is so different batches can be executed
concurrently. The dataloader runner only executes concurrently batches
that are in the same source. Having 5 different sources in some cases
slows things down because things that can naturally run in parallel
(calls to 3 different databases and parity) run sequentially.