Rework tests to use JWT auth connection with SAN tokens
After the time restriction middleware changes many queries that
previously were allowed are now not allowed. Mostly these were tests
where an error response was mocked. The to datetime was before from
datetime in the query but because the response was mocked and the
arguments ignored this was not a problem