Increase the dataloader timeout to 11 seconds.
After the increase of influxdb timeout from 5 to 10  seconds the
dataloder started to raise errors as it still had 5 seconds timeout.
The default get_policy is raise_on_error and the query failed.
Increasing to 1s more than influxdb will let influxdb timeout and not
cause an error but return an error tuple

In the future the get_policy should be replaced with `tuples` and be
handleded better