Fix timescale tests to accomodate the change with start of week day
We're updating from timescale 0.x to 1.3. From the Changelog of 1.0.0:
To better align with the ISO standard so that time bucketing starts each
week by default on a Monday (rather than Saturday), the time_bucket
epoch/origin has been changed from January 1, 2000 to January 3, 2000.
The function now includes an origin parameter that can be used to adjust this.