AccessControl middleware should be after prometheus.
AccessControl must be after prometheus because it can make a query
:resolved and that will fail the other middlewares

Also introduce a fallback to free subscriptions for cases where user
does not have subscription or user is not logged in. It is after a
catch all function that has to be removed in order to active this
fallback behavior. Doing this will make some queries not accessible at
all for users without pro or premium subscription