Recreate Influx DB on every test that uses it
From https://github.com/influxdata/influxdb/issues/10306

Dropping a measurement requires a full re-compaction of the shard(s)
containing that measurement. It isn't expected to be a fast operation as
it requires a rewrite of large portions of the persisted database.

I have tried also with dropping all the series but the behaviour was not
consistent and it frequently had timeouts.

Also re-creating the DB on every test was most performant of all the
tested methods.