Add the resolution.source to the cache key
This fixes issues where custom resolvers depend on the return value of
the top-level resolver. Example for this is the timeSeries custom
resolver for the getMetric query. It was not using the 'metric' for the
cache key and all metrics with the same datetime params were cached
under the same key