Rework how queries and metrics are handled in the access control
Instead of wrapping the metrics for getMetric in a {:metric, X} tuple
now the queries are also wraped in a {:query, Q} tuple. This will make
easier filtering and knowing if you should map &elem(&1, 1) function.
Additionally, this removes the need to filter for atom/string in order
to determine if it's a query or a metric