Handle race conditions during assert/refute called
Seems like Mock stores the called functions in a global store, so using
the same metric can lead to issues. Trying to also provide a random slug
won't work either as some functions transform the slug to something else
before passing it to the Metric module. The issue is handled by
introducing an atomic integer counter so we explicitly give a different
metric to every test - no 2 tests can run the same metric now (if the
number of tests is smaller than the number of metrics)