Reorder the tags to match the order they were when created
Issue is that we need to order them by the id in the join through
posts_tags table. Instead, when queried through the many_to_many
association they are ordered by the tag id. Example: if you create a
post with first tag SAN and second BTC due to the fact that the tag BTC
has lower id it will always come before SAN when fetching it