Rework how metric data is fetched in metric signal
The data is fetched now with 2 timeseries_data calls. This is so the
datetime rounding errors can be ignored. This happened when querying
more than 1 interval of 5-minute data with daily interval - the rounding
made is to some of the intervals were not including the proper data due
to rounding to midnight.

This is not the case with the new approach as every query is including
exactly the needed timeframes