Improve complexity calculation accuracy
In case of batching, only one of the metric names ends up in the process
dicionary and the previous ones are overwritten. Use the third argument
to deduce the complexity, if possible. In case of calling the complexity
function from timeseries_data_complexity, the third argument is the
resolution struct, which contains the metric name. In case of calling
the complexity function from the complexity phase, this struct is the
Absinthe.Complexity struct and the metric name is extracted from the
process dicionary. This works as expected in the case where one request
contains one metric but fails in some cases of multiple metrics with
different complexity weights