Handle atom values cases in the dynamic watchlist and support dynamic datetimes
- When watchlist arguments are fetched from the DB some values are
transformed from atoms to strings as json does not have the notion of
atom.

- Watchlist settings are stored in DB and are reused. This means that
static from/to datetimes are not going to work. For this, it now
supports 'dynamic_from|to' which accepts either an interval (1d) or the
string 'now'. This data is used to rewrite the from/to dynamically