Introduce labl metrics.
- Support 'label' and 'owner' fields in the metric selector
- Rework Metric module to have kw opts parameter instead of aggregation.
This allows to pass 'additional_filters' list that will be applied for
label metrics.
- Remove `exchangeVolume` GraphQL query after checking the api usage and
seeing it is not used. Now it as available as exchange
inflow/outflow/balance per exchange labeled metric and is precomputed.
- Rework the exchangeWallets and allExchanges APIs - now they do not
return infrastructure and have a mandatory 'slug' argument, so they can
return data only for a single blockchain.