Rework metric queries.
'argMax + GROUP BY asset_id, dt' leads to wrong results when
querying data from the label tables as this completely ignores the
differences in the owner and label columns. To solve this issue migrate
to using FINAL withtout argMax and GROUP BY. This can be a little bit
slower but is guaranteed to work across all cases we have.