Remove the caching from allProjectsByFunction in some cases
If all_projects_by_function is used to paginate watchlists, it uses the
base_projects field to supply the watchlist. As watchlists do not have
caching, we want to disable the caching in this case, too. This is done
to improve the UX, as otherwise users will not see any changes to the
watchlist minutes after they are made