Restructure the event bus event only when inserting in kafka
The Event structure cannot be changed itself, we can only change the
format we store in Kafka. Add configurable kafka exporter fields so in
test the max buffering messages and intervals can be reduced. This will
make the tests run faster as we don't have to wait for more
time/messages to arrive in order to push the data to kafka.