Precompute the autogenerated social volume query
The autogenerated social volume query computation has O(n^2) complexity
which results in ~4 million operations in order to compute them for all
projects. This causes timeouts when the value is not cached. In order to
avoid computing it on the fly, it is computed in the background every 15
minutes and stored in the database.