Fix wallet_movement signal and raise if such errors occur again
Before the change, the ^address pin did not match becuase of different
casing and the returned result was neither ignored nor properly handled
by the transformer. This resulted in a total breakdown of the
wallet_movement signal as the populate_payload function was called in
such a way (outside the separate task per alert) that it breaks the
execution for all alerts.