Handle Cryptocompare REST API rate limits
The rate limits are handled by pausing the whole queue when a rate limit
is found. Currently the Cryptocmpare API transports rate limit errors
over HTTP 200 (instead of 429) so the X-RateLimit-Remaining-All HTTP header
must be used to find if any of the values is 0. If it is zero, find the
value of the corresponing window in the X-RateLimit-Reset-All and use it
to resume the queue. The resuming of the queue is done by using a
separate Oban queue just for this and schedule a task that will resume
the queue after the time passes.