Fix not updating the autogenerated queries
on_conflict: :nothing has been used, which makes it impossible to update
the autogenerated queries for projects that already have some record. In
some cases no autogenerated social volume query is present in the DB. In
other cases even if the autogenerated query changed and a changeset for
this was created, it still cannot be updated. Fix by changing the
:on_conflict to :replace