Do not use a common process to send data to kafka in cryptocompare historical
Add a function to KafkaExporter that sends the data to a topic
immediatelly. Prefer using this new function instead of calling
SanExporterEx directly as the latter does not have a in-memory producer
to be used in dev/test.

This new function is used when exporting data in sync from the caller process,
so the errors can be handled and the job can be properly stored along with
the errors. This is an improvement as the Oban queue is running with a
concurrency of 25, so not sequencing everyhing through the same process
makes the processes not wait for the exporter's gen_server to handle their export
so not sequencing everyhing through the same process makes the
processes. This also makes the oban job performs more isolated as now
they do not rely on a KafkaExporter gen_server exporter running.