Rework MarkExchanges to work by using the already added labels
The old MarkExchanges worked by keeping in memory some of the exchange
addresses. The exchanges addresses are much more than we did keep in
memory (so the pod is not killed with OOM error) that this made the
marking pretty inaccurate. By using the already added labels all these
issues are fixed.