First check for deprecation, then metric availability
If a metric is deprecated it is no longer in the list of available
metrics. If the availabilty check is first, the error will not be that
the metric is deprecated, but that it is just not supported. To improve
this error message first check for deprecation. If the metric is not
supported or mistyped, the Metric.is_not_deprecated/1 will return true
and the next availability check will fail