Improve the insights searching capabilities
Introduce a prefix searching for the whole document instead of using
ILIKE only for the title. If the input is a single alphanumeric word,
treat it as an incomplete word and append ':*' to the end, so it is
searching for prefixes. For example 'mvr' would yield results very
similar to 'mvrv'.