Fix issue with locking in api calls counting (#3148)
The locking was using the whole User object which is both slower and
wrong. It is wrong because if the user changes, the lock will not work.
This can happen automatically as well - when the san balance is updated
we store the datetime value and it will cause the structs to differ.