Fix user channels (#3167)
Dispatching to several channels with the same prefix seems to not work
as intended. Attempts to join users:common resulted in trying to join
users:<user_id> which fails. Rework by joining both channels in the same
module. A special guard is added in case there is a requirement that
some message is available only to the user with the same id as the
channel name.