Entities API support multiple types (#3188)
* Improve/simplify Post sql queries

* Support :current_user_data_only param in the Entity module

This parameter allows to fetch the most voted/recent entities that are
created by the querying user. This also allows to fetch the user's
private entities

* Working most voted multiple types

* Entity.get_most_voted supports single types or list of types

* Remove timeline_event from the entities APIs

It is removed as this API will serve the purpose of fetching user
generated content. The timeline events are notifications and are not
explicilty created by the users.

* Working getMostRecent/getMostVoted for multiple types

* Add most voted entities test for multiple entities

* Fix and add test for getMostRecent for multiple types

* Refactor and add extensive docs to the Entity module

* Add spec and docs to the paginate function in the Entity module

* Fix timeline event voting tests