Move the LIMIT of top holders metrics to the top level (#3491)
When specific type of top holders - exchange or non_exchange - are
queried, it can happen that there are no exchange/non_exchange top
holders in the specified top N limit. The limit should be applied after
the JOIN so the filtering can be properly done.

If we query amount_in_exchange_top_holders with limit: 10 and the first
exchange is with rank 11, the old code won't work, but the new one will.