Fix get_metric's metadata caching (#3496)
Some fields like restricted_from and restricted_to need to be different
for users with different subscription plans. In order to implement this
without completely removing the caching, introduce the
add support for the :include_subscription_in_key flag in the
cache_resolve/1,2 macro. When set to true, the current user's plan and
product used will be included in the cache key