Rework the user registration progress state.
Instead of using a single boolean field and a single virtual boolean
field to track registration progress, now use a JSON map with many
possible states.

Fix issue where firstLogin is not properly reported for ethLogin and
google/twitter oauth login. In the case of oauth the firstLogin: true is
seen on the first currentUser query