Add firstLogin test for oauth login (#3623)
Check that after the state is in `waiting_for_login_to_finish`, which is
achieved by google/twitter oauth actions, the registration process is
finished when the `firstLogin` field is queried. This is so the frontend
does not need to introduce any locks and waits to make sure the request
that checks for firstLogin executes first and before any other currentUser
request.