Rework getMetric's metric deprecation logic (#3921)
* Rework getMetric's metric deprecation logic

Now deprecate_since JSON field can contain future dates. In such case,
the metric is accessible until that data.

* Add is_deprecated and hard_deprecate_after.

`hard_deprecate_after` replaces deprecated_since in the JSON files as it
now can contain future dates.
`is_deprecated` is also a valid boolean JSON field that acts as a soft
deprecation -- it will only mark the metric as deprecated, but it won't
hard deprecate it yet.

Both fields are exposed in the getMetric's metadata, with the same
names.