Address credo issues. Do not populate the token field in auth resolvers
Prepare the `token` return param in
ethLogin/emailLoginVerify/emailChangeVerify responses. The `token` is
the old name of `accessToken`, so both fields are duplicated. The
frontend refers to this `token` in one place, but does not use it. After
it is removed, we can remove the `token` also from the `:login` type