Address credo warnings (#4365)
* Replace is_<something> with <something>?

* Address credo code duplication reports

* Address credo code complexity reports

* Address credo warnings regarding complex functions

* Address credo issues. Do not populate the token field in auth resolvers

Prepare the `token` return param in
ethLogin/emailLoginVerify/emailChangeVerify responses. The `token` is
the old name of `accessToken`, so both fields are duplicated. The
frontend refers to this `token` in one place, but does not use it. After
it is removed, we can remove the `token` also from the `:login` type

* Eliminate shift-reduce conflict in san_lang_parser.yrl

`value` can be derived from `expr`.

* Fix tests using token instead of accessToken