Fix ueberauth callback failure handling (#4474)
We provide a custom `request/2` function, so in the callback we manually
invoke Ueberauth.run_callback/2. Because of this when we enter the
function, the assigns field of the conn is not yet filled, so we don't
know if it is a failure or not. Because of this we cannot handle failure
with callback(%{assigns: %{ueberauth_failure: _}}) but we need to do it
after invoking Ueberauth.run_callback