Add timeseries data JSON result queries
Returning list of objects results in excessive type checking and boxing
done by Absinthe. In the end, it can happen that a response that is less
than 1Mb in JSON format requires more than 50Mb of RAM memory to process
because it creates an in-memory struct for each value in the result.

Introduce *Json queries which do exactly the same thing as the original
ones, but only change the result type to JSON. This makes it impossible
to specify which fields you want to select, but it requires much less
RAM memory when running. A quick test showed that in the original
queries when the size of the response goes over a few megabytes, it can
consume hundreds megabytes of RAM just to process. In the new JSON
queries we can easily return results that are 25Mb in JSON.