Filter out stories that are not ready yet
Trending stories are computed in 2 phases. After the initial insertion
in the DB, some of the fields are empty. They are filled (if there is
story found) on a later step. This causes the API to not work properly
for a few minutes after round hours -- the fields are just empty.