Solve the PricePair renaming of slugs in result
Cryptocompare data is stored with their cryptocompare base_asset name.
Santiment has multiple assets pointing to the same cryptocompare asset
like: arb-arbitrum -> ARB, arbitrum -> ARB, etc.

In the SQL we first map the san asset to the cryptocompare asset, and
then we map the cryptocompare asset to the san asset (in SQL). In this
second operation we lose track of the original asset, so we end up in a
situation where we map arb-arbitrum to ARB, but then we map ARB to
arbitrum, in practice renaming the slug in the result.