Introduce more flexible internals for getMost* APIs
Previously we had to match on the cartesian product of two variables
({true, false}, {false, true} and {true, true}) to know whether to fetch
all public entities, user private entities or public + user entities.

Now that we needed to extend the behavior a bit, introducing a new
variable would have required to exponentially increase the number of
matches, becoming unmaintainable.

Rework the internals so now everything depends on the opts and one SQL
query which is dynamically built based on those opts